Corrected Critter behaviour
Now Power Critter can fly! Also he makes the screen shake when he lands.

The o->blocku stuff was silly. If a critter touches the block above him
while flying, he lands, so why the hell should doing so make him start
flying? Even after enable Power Critter to fly, he didn't seems to want to
fly most of the time. It was because of that.