Strafing support like Switch CS+
A bit of a feature copy - Cave Story+ on the Nintendo Switch has a "feature"
where by holding either ZL or ZR Quote will not change direction while moving,
allowing one to backstep while firing the weapon at an enemy or lock vertical
trajectory to one side with the fireball.

This adds a new key (KEYSTRAFE) and makes it so Quote's direction/look
is not changed when it is held. It also adds a setting to make this
added feature toggleable, since strafing makes some parts far easier.

Unfortunately, this breaks the settings file since a new key was added
as well as saved option for the toggle. The version was incremented
(NSX3 -> NSX4) to avoid problems.