Update inventory position every frame
If the user switches to/from widescreen mode while the inventory is
open, the position of the inventory will now update instead of being
offset one way or the other. Also, it's not necessary to adjust the y
position of the inventory based on whether we're in widescreen mode
because the screen height doesn't change.