Fix uninitialized memory error
Renderer::initVideo now depends on Sprites being initialized because it
calls setResolution. If the sprites aren't initialized, the Renderer can
get a segfault in flushAll() when it tries to check which spritesheets
are initialized and which aren't. Sprites can only detect whether the
spritesheets are initialized after Sprites::init has been called, which
zeroes out the spritesheets.

This bug only appeared on my Raspberry Pi, so I didn't notice it in the
last several months of testing. Presumably the other systems I tested on
had this uninitialized memory set to all zeroes by chance.