Don't show the Sue title after completing the game without niku counter
If you finish Sacred Grounds without the nikumaru counter, the game
saves "0" in 290.rec. This used to have two consequences:

1. The title screen would change to the Sue screen, the best possible.
2. If you ever beat Sacred Grounds with the nikumaru counter, your new
   time would never be better than 0 and would not be saved.

Now if we load 290.rec and see that it's 0, we don't take that value
literally. We assume 0 to mean that no time has been set. This fixes
both issues.

Fixes #255